src/benchmark/scalability benchmarks/requirements.txt
kafka-python>=2.0.2
matplotlib>=3.5.0
seaborn>=0.12.0
numpy>=1.21.0
requests>=2.28.0
